use new example structure
