--- packages/example/README.md
@@ -1,1 +1,0 @@
-This project was bootstrapped with [Create React App](https://github.com/facebookincubator/create-react-app).
