--- packages/example/not-package.json
@@ -1,40 +1,0 @@
-{
-  "name": "example",
-  "version": "0.0.1",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/helloeave/flow-form.git"
-  },
-  "scripts": {
-    "build": "echo ignoring: react-scripts build",
-    "eject": "react-scripts eject",
-    "start": "PORT=1337 react-scripts start",
-    "test": "react-scripts test --env=jsdom"
-  },
-  "dependencies": {
-    "@flow-form/field": "^0.0.1",
-    "@flow-form/form": "^0.0.1",
-    "@flow-form/slider": "^0.0.1",
-    "history": "^4.7.2",
-    "libphonenumber-js": "^1.0.3",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.0",
-    "react-syntax-highlighter": "^6.1.2"
-  },
-  "devDependencies": {
-    "eslint": "^4.9.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.4.0"
-  },
-  "browserslist": {
-    "development": ["last 2 chrome versions", "last 2 firefox versions", "last 2 edge versions"],
-    "production": [">1%", "last 4 versions", "Firefox ESR", "not ie < 11"]
-  }
-}
