+++ packages/example/package.json
@@ -7,14 +7,20 @@
   "private": true,
   "scripts": {
     "build": "NODE_ENV=production webpack --config ./webpack.config.js --mode production",
-    "dev":
-      "NODE_ENV=development webpack-dev-server --mode development --hot --history-api-fallback",
+    "dev": "NODE_ENV=development webpack-dev-server --mode development --hot --history-api-fallback",
     "analyze": "ANALYZE=true yarn run build"
   },
   "dependencies": {
+    "@flow-form/extra-fields": "0.0.1",
+    "@flow-form/field": "0.0.1",
+    "@flow-form/form": "0.0.1",
+    "@flow-form/helpers": "0.0.1",
+    "@flow-form/slider": "0.0.1",
     "foundation-sites": "^6.4.4-rc1",
     "history": "^4.7.2",
+    "libphonenumber-js": "^1.1.6",
     "lodash": "^4.17.5",
+    "mini-css-extract-plugin": "^0.3.0",
     "prop-types": "^15.6.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
@@ -22,6 +28,7 @@
     "react-redux": "^5.0.7",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
+    "react-syntax-highlighter": "^7.0.2",
     "redux": "^3.7.2"
   },
   "devDependencies": {
@@ -31,6 +38,7 @@
     "autodll-webpack-plugin": "^0.3.9",
     "autoprefixer": "^8.2.0",
     "babel-loader": "^7.1.4",
+    "babel-plugin-lodash": "^3.3.2",
     "babel-plugin-syntax-decorators": "^6.13.0",
     "babel-plugin-transform-decorators": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
@@ -44,11 +52,13 @@
     "html-webpack-plugin": "^3.1.0",
     "mini-css-extract-plugin": "^0.3.0",
     "node-sass": "^4.8.3",
+    "optimize-css-assets-webpack-plugin": "^4.0.0",
     "postcss": "^6.0.21",
     "postcss-loader": "^2.1.3",
     "sass-loader": "^6.0.7",
     "style-loader": "^0.20.3",
     "thread-loader": "^1.1.5",
+    "uglifyjs-webpack-plugin": "^1.2.4",
     "webpack": "^4.3.0",
     "webpack-bundle-analyzer": "^2.11.1",
     "webpack-cli": "^2.0.13",
