+++ packages/example/src/components/NavBar.js
