+++ packages/example/src/components/NavBar.scss
