+++ packages/example/src/components/pages/DependentField.js
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
+import { hot } from 'react-hot-loader';
 
+@hot(module)
 export default class DependentField extends Component {
   render() {
     return (
