+++ packages/example/src/components/pages/Fields.js
@@ -1,0 +1,5 @@
+import React, { PureComponent } from 'react';
+import { hot } from 'react-hot-loader';
+
+@hot(module)
+export default class Fields extends PureComponent {}
