--- packages/example/src/index.css
@@ -1,39 +1,0 @@
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-family: Montserrat, sans-serif;
-}
-
-body {
-  margin: 0;
-  padding: 0;
-  font-family: 'Open Sans', sans-serif;
-}
-
-.remove-fieldset-styling {
-  border: 0;
-  padding: 0;
-  margin: 0;
-  min-width: 0;
-}
-.remove-fieldset-styling legend {
-  margin: 0;
-  padding: 0;
-}
-
-pre,
-code {
-  font-family: monospace;
-  background-color: #eee;
-}
-
-code {
-  padding: 0 0.25em;
-}
-
-pre {
-  padding: 1em;
-}
