+++ packages/example/src/index.js
@@ -1,7 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { render } from 'react-dom';
+import App from './app';
 
-import './index.css';
-import App from './App';
+require('./index.scss');
 
-ReactDOM.render(<App />, document.getElementById('root'));
+render(App, document.getElementById('app'));
