+++ packages/example/src/store.js
