+++ packages/form/package.json
@@ -64,5 +64,6 @@
     "prop-types": "^15.6.1",
     "react": "^15.0.0 || ^16.0.0",
     "react-dom": "^15.0.0 || ^16.0.0"
-  }
+  },
+  "sideEffects": false
 }
