--- packages/new-example/src/components/pages/Intro.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Intro() {
-  return <div>This is the intro.</div>;
-}
