--- packages/new-example/src/index.js
@@ -1,6 +1,0 @@
-import { render } from 'react-dom';
-import App from './app';
-
-require('./index.scss');
-
-render(App, document.getElementById('app'));
