+++ packages/slider/package.json
@@ -70,5 +70,6 @@
   },
   "publishConfig": {
     "access": "public"
-  }
+  },
+  "sideEffects": false
 }
